move tokens to api
