+++ src/stores/tokens.js
@@ -1,12 +1,10 @@
 import { computed, ref } from 'vue'
 import { useRuntimeConfig } from 'nuxt/app'
 import { defineStore } from 'pinia'
-import useAxios from '@/composables/useAxios'
 import { adaptListedTokens } from '@/utils/adapters'
 
 export const useTokensStore = defineStore('tokens', () => {
-  const { MIDDLEWARE_URL, DEX_BACKEND_URL, NETWORK_NAME } = useRuntimeConfig().public
-  const axios = useAxios()
+  const {  NETWORK_NAME } = useRuntimeConfig().public
   const rawListedTokens = ref(null)
   const allTokens = ref(null)
   const selectedTokenName = ref(null)
@@ -25,31 +23,21 @@ export const useTokensStore = defineStore('tokens', () => {
       : null,
   )
 
-  function fetchTokens(queryParameters) {
-    return Promise.allSettled([
-      fetchAllTokens(queryParameters),
-      fetchListedTokens(),
-    ])
+ async  function fetchTokens(queryParameters) {
+   allTokens.value = null
+   rawListedTokens.value = null
+   // test mainnet testnet
+// todo pass query params
+
+   // todo rename
+    const aaa = await $fetch('/api/tokens')
+    allTokens.value = aaa[0].value
+    rawListedTokens.value = aaa[1].value
   }
 
-  async function fetchAllTokens(queryParameters = null) {
-    allTokens.value = null
-    const { data } = await axios.get(`${MIDDLEWARE_URL}${queryParameters || '/v3/aex9?by=name&direction=forward'}`)
-    allTokens.value = data
-  }
-
-  async function fetchListedTokens() {
-    rawListedTokens.value = null
-    const { data } = await axios.get(`${DEX_BACKEND_URL}/tokens/listed`)
-    rawListedTokens.value = data
-  }
 
   async function fetchTokensCount() {
-
-    allTokensCount.value =
-      await $fetch('/api/tokens/count')
-
-    allTokensCount.value = data
+    allTokensCount.value = await $fetch('/api/tokens/count')
   }
 
   return {
